refactor: last state들 제거
