+++ src/main/java/bowling/domain/frame/CommonFrame.java
@@ -1,18 +1,22 @@
 package bowling.domain.frame;
 
+import bowling.domain.dto.StateData;
+import bowling.domain.pin.Pins;
 import bowling.domain.score.Score;
-import bowling.domain.state.CommonState;
+import bowling.domain.state.State;
 import bowling.domain.state.Start;
 
 import java.util.List;
 
 public class CommonFrame extends Frame {
+
     private static final int COUNT_OF_COMMON_FRAME = 9;
 
     private Frame nextFrame;
+    private State state;
 
-    private CommonFrame(CommonState state) {
-        super(state);
+    private CommonFrame(State state) {
+        this.state = state;
         nextFrame = DummyFrame.of();
     }
 
@@ -22,7 +26,15 @@ public class CommonFrame extends Frame {
 
     @Override
     public Score getScore() {
-        return nextFrame.addBonusScore(state.score());
+        return nextFrame.addStateScore(state.score());
+    }
+
+    public void hitPins(Pins pins) {
+        this.state = state.hitPins(pins);
+    }
+
+    public StateData getFrameStates() {
+        return StateData.of(state.getState());
     }
 
     @Override
@@ -35,10 +47,10 @@ public class CommonFrame extends Frame {
     }
 
     @Override
-    protected Score addBonusScore(Score score) {
+    protected Score addStateScore(Score score) {
         Score addedScore = state.addScore(score);
 
-        return nextFrame.addBonusScore(addedScore);
+        return nextFrame.addStateScore(addedScore);
     }
 
     private boolean isFrameNotFinish() {
@@ -57,6 +69,41 @@ public class CommonFrame extends Frame {
     private void updateNextFrame(List<Frame> frames, Frame nextFrame) {
         this.nextFrame = nextFrame;
         frames.add(nextFrame);
+    }
+
+    private static class DummyFrame extends Frame {
+
+        private State state;
+
+        private DummyFrame(State state) {
+            this.state = state;
+        }
+
+        public static DummyFrame of() {
+            return new DummyFrame(Start.of());
+        }
+
+        @Override
+        public Score getScore() {
+            return Score.NULL;
+        }
+
+        @Override
+        protected Score addStateScore(Score score) {
+            return score;
+        }
+
+
+        @Override
+        protected void hitPins(Pins pins) {
+            this.state = state.hitPins(pins);
+        }
+
+        @Override
+        protected StateData getFrameStates() {
+            return StateData.of(state.getState());
+        }
+
     }
 
 }
