+++ src/main/java/bowling/domain/state/Strike.java
@@ -1,5 +1,6 @@
 package bowling.domain.state;
 
+import bowling.domain.pin.Pins;
 import bowling.domain.score.ProgressScore;
 import bowling.domain.score.CommonScore;
 import bowling.domain.score.Score;
@@ -9,7 +10,7 @@ import java.util.List;
 
 import static bowling.domain.pin.Pins.MAX_COUNT_HIT_PINS;
 
-public class Strike extends End {
+public class Strike extends State {
 
     public static Strike of() {
         return new Strike();
@@ -33,6 +34,16 @@ public class Strike extends End {
     @Override
     public Score addBonusScore(Score score) {
         return score.add(CommonScore.ofStrike());
+    }
+
+    @Override
+    public boolean isFinish() {
+        return true;
+    }
+
+    @Override
+    public State hitPins(Pins pins) {
+        throw new IllegalStateException("다음은 없습니다");
     }
 
 }
